Added password validation to adding employees on the desktop, as well as the password requirements when creating new employees.
The password validation is done in the new controller for the AddEmployeeForm and it goes through the password char by char to see if it contains the necessary characters.

The UI for adding employees now shows password requirements and shows if a password is valid or invalid when typing text.

The submit button now is disabled untill the password is valid.