Progress on viewing rentals and renting/returning them on desktop.
UI is done, as well as the empty shell methods needed in the view.

Some working code is done but this is not using the DbContext so work there will be moved to using the DbContext.

Added a new Database class on desktop, for now it has functionality but it will be removed with the transistion to the DbContext.