Implemented using DbContext for logging in, viewing and managing employees, adding employees, and viewing stock for Desktop.
UI is there for rentals but SQL is not fully implemented.

Using the DbContext simplified code a lot and I was able to remove a lot of lines of code and all of the Database and Model code inside of the Desktop project. The DbContext has some qeird quirks such as having the forign keys has a item attached to that DbContext model, because of this items currently have a ugly default tostring message.

Removed uneeded tests.