[Desktop] Made all forms but the login form open and close for proper logout functionality.
This is so that the flow of how going between forms is better, and that logging out now works correctly.

Changes were made in ManageEmployeeForm, ManageItemsForm, and ManageRentalsForm...

Change consists of going from using Hide -> Show to now doing Hide -> Close when the form is returned to. (basically creating a chain of closures for the logout and exiting the form from the X button.

This change cannot be made on the LoginForm since closing the LoginForm will always close the entire application!