[Desktop] Hid unwanted columns from the various DataGridViews.
This change was done to ensure that the data in the grid views was clean and easy to read, as the unwanted columns were just ugly toStrings with no real purpose.

This change was acomplished by adding a Columns[x].Visiable = false for each of the unwanted columns inside of ManageEmployeeForm, ManageItemsForm, and ManageRentalsForm.

Removed some old inline comments as well to keep code clean.