[Desktop] Removed ID from add employee and moved to using the stored procedure for inserting an employee.
The ID was removed from adding an employee so that the web and desktop both use the auto increment feature of the MySql table for employees.

The swap to using the stored procedure allows us to unify the encryption and decription of passwords using the built in Sql encrypt.

The ID lable and numericUpDown were removed from AddEmployeeForm, and all mentions of it in code removed as well.

The attempt at building an Employee in the AddEmployeeForm code behind was removed as the stored procedure simply takes the name, manager status, and password.