[Desktop] Added verification to removing an employee.
This is a quality of life improvement that is generally expected when forming a irreversable task on a dataset.

This was implemented by adding a new dialog form called RemoveConfirmationForm, the form takes the employee to remove as a parameter and asks if the user wants to remove it by name with yes or no options. This form has relitevily little code and simply returns the users selection as a DialogResult.

If the employee is null it returns an abort message which can be displayed to the user.  Selecting No does nothing as the selection uses a switch case to determine the action.